Remove entity decoder code that is never called (#7)
As of html.parser version 3.5, the handle_charref() method never gets called. This is because HTMLParser.__init__() now sets a flag to automatically convert (i.e. decode) character_refs and entity_refs. This flag is configurable, but by default is set to true.

From the html.parser docs:
"Changed in version 3.5: The default value for argument convert_charrefs is now True."
"If convert_charrefs is True (the default), all character references (except the ones in script/style elements) are automatically converted to the corresponding Unicode characters."

This commit updates the call to HTMLParser.__init__ to be compatible with v3.5.
Note that if convert_charrefs is set to true, HTMLParser calls unescape() itself when handling data. Therefore, overriding handle_charref() is no longer required.